+++ lizard_portal/media/lizard_portal/lizard/grid/ComboDict.js
@@ -104,8 +104,4 @@ Ext.define('Lizard.grid.ComboDict', {
             return this.getValue();
         }
     }
-
-
-
-       
 });