+++ lizard_portal/media/lizard_portal/lizard/grid/EditableGrid.js
@@ -22,7 +22,41 @@
     - multiselect comboboxes doesn't work correct
   
   
-  */  Ext.define('Lizard.grid.EditableGrid', {
+  */  Ext.override(Ext.form.field.Field, {
+    isEqual: function(a, b) {
+      if (Ext.isDate(a) && Ext.isDate(b)) {
+        return a.getTime() === b.getTime();
+      } else if (Ext.isObject(a) && Ext.isObject(b)) {
+        return a.id === b.id;
+      } else if (Ext.isArray(a) && Ext.isArray(b)) {
+        console.log(a);
+        console.log(b);
+        if (a.length === 1 && b.length === 1) {
+          console.log(a[0].id === b[0].id);
+          return a[0].id === b[0].id;
+        }
+      }
+      return a === b;
+    }
+  });
+  Ext.override(Ext.data.Model, {
+    isEqual: function(a, b) {
+      if (Ext.isDate(a) && Ext.isDate(b)) {
+        return a.getTime() === b.getTime();
+      } else if (Ext.isObject(a) && Ext.isObject(b)) {
+        return a.id === b.id;
+      } else if (Ext.isArray(a) && Ext.isArray(b)) {
+        console.log(a);
+        console.log(b);
+        if (a.length === 1 && b.length === 1) {
+          console.log(a[0].id === b[0].id);
+          return a[0].id === b[0].id;
+        }
+      }
+      return a === b;
+    }
+  });
+  Ext.define('Lizard.grid.EditableGrid', {
     extend: 'Ext.grid.Panel',
     alias: 'widget.leditgrid',
     config: {
@@ -119,6 +153,7 @@
           editor = me.editors[type];
         } else if (type === 'combo') {
           if (col.choices && col.choices.length > 0 && Ext.type(col.choices[0]) === 'object') {
+            console.log('combodict');
             editor = {
               field: {
                 xtype: 'combodict',
@@ -137,6 +172,7 @@
               }
             };
           } else if (col.remote) {
+            console.log('gridcombobox');
             editor = {
               field: {
                 xtype: 'gridcombobox',
@@ -156,6 +192,7 @@
               }
             };
           } else {
+            console.log('combo');
             editor = {
               field: {
                 xtype: 'combo',
@@ -339,7 +376,6 @@
           cols.push(cols_with_header);
         }
       }
-      console.log(cols);
       return cols;
     },
     saveEdits: function() {
@@ -421,14 +457,16 @@
           extraParams: {},
           reader: {
             type: 'json',
-            root: 'data'
+            root: 'data',
+            totalProperty: 'count'
           },
           writer: {
             type: 'json',
             writeAllFields: false,
             root: 'data',
             encode: true,
-            successProperty: 'success'
+            successProperty: 'success',
+            totalProperty: 'count'
           },
           autoLoad: true
         }
@@ -439,7 +477,8 @@
       var me;
       me = this;
       me.columns = this.getColumnConfig();
-      me.store = this.getStoreConfig();
+      me.store = Ext.create('Lizard.store.EditGridStore', this.getStoreConfig());
+      me.bbar = [];
       if (this.getEditable()) {
         this.editing = Ext.create('Lizard.grid.CellEditing', {
           clicksToEdit: 1
@@ -500,13 +539,13 @@
           }
         ]);
       }
-      this.on('edit', function(editor, e) {
-        console.log('editor:');
-        console.log(editor);
-        console.log('e:');
-        console.log(e);
-        return console.log(editor.getActiveEditor());
-      });
+      this.bbar = {
+        xtype: 'pagingtoolbar',
+        pageSize: 25,
+        store: me.store,
+        displayInfo: true,
+        items: ['-'].concat(me.bbar)
+      };
       return this.callParent(arguments);
     }
   });